Bugfix for use of __FILE__ macro
The __FILE__ macro was being placed in quotes when assigned to a filename
parameter. This is incorrect because it is already a quoted string. Remove the
quotes. change the filename variable to mod_filename to be more clear.

Test suite: compiled and run with dummy input. now prints filename correctly.