Read namelist for simulator and reaction class names.
Driver reads the names of the simulator and reactions classes from a
namelist instead of hard coded. Fix a bug in the select case in the
simulator factory (was specifying reaction name instead of a simulator
name). Update namelist names to be consistent with code.

Test status: compiles and run on os x. reads namelist correctly.
             segfaults during standalone init with segfault.