Initial implementation of BeTRSimulationCLM
Start working on the implementation of BeTRSimulationCLM by copying
BeTRSimulationStandalone and moving code from the betr_clm_cpl module.
Note that there is duplicate code between the standalone simulation and
clm simulation. It's not clear yet what is necessary and what can be
removed by refactoring. BeTRSimulationCLM is not being compiled in this
change set, and will not compile if added to the build.

Testing: os x configure, build test pass
         standalone mock - no difference in results.