Initial working implementation of BeTRSimulationCLM
Initial working implementation of the BeTRSimulationCLM class. Contains
a lot of duplicate code that will be removed when consolidating the
simulation classes. Several functions are called by clm and will
eventually be removed and put into the clm side of the interface.

Testing: mac os x config, build, test ok.
         runs standalone mock and clm mock to completion, no numerical
           differences with reference output.