Move initialization of the betr lsm coupler into betr class
Move the initialization of the betr-lsm coupler,
init_betr_lsm_bgc_coupler, out of the simulation class into the betr
class.

Note, ecophyscon is part of the function arguments, but does not appear
to be used or even initialized properly. We are raising an error if it
used and passing junk.

Testing: os x configure, build, test - pass
         clm-mock - bit for bit
         clm-h2oiso - expected runtime mass balance error.