Refactor and unit tests for regression test driver
Move the comparison code out of the RegressionTest class into a separate
Comparison class so it is more clear where the comparison code is.

Write unit tests for the Tolerances and Comparison classes. Some
logic/algorithm tests in Tolerance, but mainly verifying that we are
correctly trapping on invalid input. Basic test that valid and invalid
floating point comparison with absolute tolerance is working as expected.

Save a previous test log to a backup file.

Testing: os x gfortran configure, build, unit tests - pass
         regression tests - mock pass, h2oiso expected failure
         meta-tests - pass