Loosen tolerances in standalone h2oiso test
Loosen the tolerances in the standalone h2oiso test to get a clean
checkout and build passing on another machine.

Failures:

    FAILURE : h2oiso : O18_H2O : min : tolerance failure : 1.00044417195e-11 > 1e-16 [absolute]
    FAILURE : h2oiso : O18_H2O : min : tolerance failure : 1.00044417195e-11 > 1e-16 [absolute]

Loosening the absolute tolerance on all the concentrations in this test
would have resulted the uncomfortably large tolerances. The actual
numerical values:

    baseline = 3.5756650343087E+02
    current = 3.5756650343088E+02

are only off by single digit off in the least significant figure.
Changing to a relative tolerance allows this to pass and preserves the
tolerance on other values.

Testing: os x configure, build - ok
         unit tests - pass
         regression tests - pass
         mtest - pass