Runtime configuration for transport processes
Add the runtime switches for turning on/off advection, diffusion and
reaction. default is on. Incorporate the switches for advection and
diffusion into the correct location in BetrBGCMod::tracer_gw_transport.
The reaction flag is not currently implemented. Add standalone mock
versions of pure advection, diffusion and reaction tests. When advection
is turned off, there is a mass balance error that generates a runtime
error.

Tests: os x gfortran - configure, build
       unit tests - pass
       regression tests - pure diffusion and reaction fail at runtime, all others pass.
       meta tests - pass