suppress all compiler warnings in clm directories
We don't want to make extensive changes to the clm stub code, but there
are significant numbers of compiler warnings in that code. Suppress all
compiler warnings in code we don't intended to change. But keep the
compiler warnings in the betr code base.

Testing: configure and build - os x, gfortran 5.3.0
         unit tests - pass
         regression tests - pass
         meta tests - pass