Improve regression test input data conversion
Improve the conversion of regression test input from text cdl to binary
nc files. Look at the modification time on the file system to determine
if the cdl file is newer and the nc file needs to be updated. Write
conversion info to log file. Generate filenames by appending '.nc' to
the end of the filename. Make clobber rule to remove all test artifacts.

Testing:
    configure and build - os x, gfortran 5.3.0
    unit tests - pass
    regression tests - pass
    meta tests - pass