Remove global from ode mod by cleaning up callbacks
Remove the global mbkks_data from ode module. Rename the struct as
func_data_type, and put it into it's own module as a generic container
to pass to all call back functions. Add the generic container to all
call back function signatures, and pass into calls. This allowed the
removal of copy and paste routines in find root mod for brent and
hybrid_findroot that only differed by the args to the call back.

Testing:
    configure and build - os x, gfortran 5.3.0
    unit tests - pass
    regression tests - pass
    meta tests - pass