Search for ncgen during regression tests
The regression test driver automatically converts input data from text
cdl to binary netcdf. Previously we assumed that ncgen would be in the
user's path. Since we are now building our own netcdf, we try that
version first, and only fall back to the version in the path if our
version doesn't work. Provide an explicit error if a useable ncgen can't
be found.

Testing:
    configure and build - os x, gfortran 5.3
    unit tests - pass
    regression tests - pass
    meta tests - pass