Decrease run length of two regression tests
Running the pure diffusion and pure reaction regression tests for 100s
of days us unnecessary. It doesn't contribute anything to science or
regression testing. The reaction test is all zeros because there are no
reactions. Diffusion profiles are similar between the shorter and longer
test. But it does increase the test turn around time and
probability of a test timing out.

Testing:
    configure and build - OS X, gfortran 5.3
    unit tests - pass
    regression tests - pass, update mock-diffusion baseline.
    meta tests - pass