pull hisotry output into simulation class
Now the history output is executed in the simulation class,
which can be easily hooked with clm/alm. By default, this
capability is turn on at the moment, because it use subrtouines
from histFileMod. It might be good to remove the dependence on HistFileMod
in later iterations, or at least partially duplicate what is done
in ALM/CLM's histFileMod

testing:
    configure and build - os x, gfortran 6.0.0
    unit tests - pass
    regression tests - pass
    meta tests - pass