removed some static memory allocation for setting history fields.
This is acheived through the introduction of a list structure, which
expands dynamically with the addition of new variables. After this
change, all tests seem to pass, with occasionally check_sum error
though. The issue may be linked to stack size limit or actually memory
leaking that is likely to occur in TracerBaseType.F90. More
exploration is needed on other platforms.
test:
   configure&build: pass OSx, gcc6.2.0_0
   mtest: pass
   unit test: pass
   regression test: pass if the check_sum error did not happen