remove a potential repetitive memory deallocation in transportmod
The extra_inst in transportmod has the possibility to free the memory twice. Removing
this potential bug however did not resolve the checksum error. Some investigation
indicates the problem occurs only when transient data used, and is usally linked to
netcdf io. Therefore, a possibility is that sbetr triggered some bug in the
netcdf library. So the next step would be to check if alm-betr would have this problem or not.
Also, it was noticed intel compiler does not trigger this problem. It is therefore
good to check with compilers other than gnu and intel.

testing:
  configure & compile: pass, OSx gcc6.2.0_0
  mtest: pass
  unit test: pass
  regression test: pass with occasional failure when transient forcing data is used.