update for betr-eca_cnp and some bug fix
Now betr-eca_cnp is running with correct mass balance
for all involved chemical elements. Also, some bug fix
is done to constrain the print to screen output to only
the master processes. This also resulted in some revision
of the betr interface for other bgc reactions that were
currently included in betr_rxns