bug fix for excessive iterations in advection
In cases of a freezing or a drying soil, the liquid water amount may be extremely
small, causing a blow up in advective velocity, which consequently leads to
too many iterations in semi-lagrangian advection. Now this is fixed by
adding a minimum liquid water amount that is not too small in computing the
phase conversion parameter for aquesous tracers. Also included are a number
of debug information subroutines.

Now the code does betr-eca-cnp simulations smoothly. Tests were done with
intel compiler on edison.